aegisk: Do VT switch manually, without requirring an attached TTY
Previously we called an ioctl on get_current_tty(), but when the
aegis usermode helper triggers lock_vt, it doesn't have a tty.
Fortunately, the fg_console and vt_dont_switch are actually
global state, and nothing that the TTY ioctl does actually requires
a TTY here.
So we can just do what the ioctl does internally, by resolving
the right symbols with KLA and using them directly.